Import and export and cleanup

+++ db/migrate/20130925021312_create_members.rb
@@ -32,7 +32,7 @@ class CreateMembers < ActiveRecord::Migration
       t.integer :blk
       t.string :side
       t.string :multi
-      t.decimal :evenodd
+      t.integer :evenodd
       t.string :on
       t.string :ref
       t.string :plus
